moved components into their parent directories and deleted additional components?
