+++ src/App.js
@@ -3,8 +3,8 @@ import { BrowserRouter, Switch, Route, Redirect } from "react-router-dom";
 import { PrivateRoute } from "./Components/Helpers/PrivateRoute";
 import Splashpage from "./Components/Splashpage";
 import Signup from "./Components/Signup";
-import ResetPassword from "./Components/ResetPassword";
-import ForgotPassword from "./Components/ForgotPassword";
+import ResetPassword from "./Components/Login/ResetPassword";
+import ForgotPassword from "./Components/Login/ForgotPassword";
 import OrgSelect from "./Components/OrgSelect";
 import Dashboard from "./Components/Dashboard";
 import BoilerplatesIndex from "./Components/Boilerplates/BoilerplatesIndex";
@@ -21,8 +21,8 @@ import FundingOrgNew from "./Components/FundingOrgs/FundingOrgNew";
 import BoilerplatesNew from "./Components/Boilerplates/BoilerplatesNew";
 import OrganizationsNew from "./Components/OrganizationsNew";
 import GrantsNew from "./Components/Grants/GrantsNew";
-import SectionsNew from "./Components/SectionsNew";
-import ReportsNew from "./Components/ReportsNew";
+import SectionsNew from "./Components/Sections/SectionsNew";
+import ReportsNew from "./Components/Reports/ReportsNew";
 import GrantEdit from "./Components/Grants/GrantEdit";
 import GrantCopy from "./Components/Grants/GrantCopy";
 import OrganizationLayout from "./Components/Layouts/OrganizationLayout/OrganizationLayout";
