+++ src/Components/Reports/ReportsShow.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
-import ReportSectionsNew from "./ReportSectionsNew";
+import ReportSectionsNew from "./ReportSections/ReportSectionsNew";
 import ReportSectionsShow from "./ReportSectionsShow";
 import Card from "react-bootstrap/Card";
 import Button from "react-bootstrap/Button";
