+++ src/Components/Sections/SectionsNew.js
@@ -4,10 +4,10 @@ import Card from "react-bootstrap/Card";
 import Button from "react-bootstrap/Button";
 import ReactQuill from "react-quill";
 import "react-quill/dist/quill.snow.css";
-import { useCurrentOrganization } from "../Contexts/currentOrganizationContext";
-import { createGrantSection } from "../Services/Organizations/Grants/GrantSectionsService";
-import { getAllBoilerplates } from "../Services/Organizations/BoilerplatesService";
-import countWords from "../Helpers/countWords";
+import { useCurrentOrganization } from "../../Contexts/currentOrganizationContext";
+import { createGrantSection } from "../../Services/Organizations/Grants/GrantSectionsService";
+import { getAllBoilerplates } from "../../Services/Organizations/BoilerplatesService";
+import countWords from "../../Helpers/countWords";
 
 export default function SectionsNew(props) {
   const [quillText, setQuillText] = useState("");
